requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.10.0
Django==1.11
django-crispy-forms==1.6.1
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
lxml==3.7.3
openpyxl==2.4.7
packaging==16.8
pycparser==2.17
PyMySQL==0.7.11
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
cloudflare==1.5.1
django_slack==5.8.0
django-extensions==1.7.9
fuzzywuzzy==0.15.0
gevent==1.2.2
